powerpc:powernv: Perf driver for inband-OCC sensors
This patch adds a new perf driver to export the OCC inband-sensors.
OCC is an on-chip-controller which controls and monitors the thermal
and power health of the chip by reading various thermal-power sensors.
OCC copies few of these sensors to main memory which will be mapped
by this driver. This driver exports the below sensors:
System level-
1)system-power 2)gpu-power 3)io-power 4)storage-power 5)fan-power
6)fan-speed 7)system-energy 8)current-pcap 9)max-pcap 10)hard-min-pcap 11 )soft-min-pcap
Chip related sensors-
1)Power 2)chip-energy 3)chip-mbw 4)power-memory 5)power-vcs
5)power-vdd
Core related sensors-
1)core-temp for all the cores in the chips.

Inband sensors have low read latency compared to other ways like
IPMI & XSCOM. This driver is an initial implementation (POC)
of inband sensor framework which can be extended for P9.

This driver is based on the below firmware changes:
1) OCC : <https://github.com/code-hippo/occ/tree/OPAC>
2) skiboot : <https://github.com/code-hippo/skiboot/tree/OPAC>

Signed-off-by: Bhargav Reddy <bhargav431994@gmail.com>